Switch to using DateTime?s for dates from MailChimp
