+++ MailChimpV3API/MailChimpV3API.csproj
@@ -47,7 +47,7 @@
     <DebugType>pdbonly</DebugType>
     <PlatformTarget>x64</PlatformTarget>
     <ErrorReport>prompt</ErrorReport>
-    <CodeAnalysisRuleSet>MinimumRecommendedRules.ruleset</CodeAnalysisRuleSet>
+    <CodeAnalysisRuleSet>..\codestyle.ruleset</CodeAnalysisRuleSet>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)' == 'Debug|x86'">
     <DebugSymbols>true</DebugSymbols>
@@ -171,16 +171,28 @@
     <Compile Include="Sections\Lists\MergeFields\MergeFieldOptions.cs" />
     <Compile Include="Sections\Lists\MergeFields\MergeFieldsCollection.cs" />
     <Compile Include="Sections\Lists\MergeFields\MergeFieldsFilter.cs" />
+    <Compile Include="Sections\Reports\ABSplit.cs" />
+    <Compile Include="Sections\Reports\ABSplitItem.cs" />
+    <Compile Include="Sections\Reports\Bounces.cs" />
+    <Compile Include="Sections\Reports\Clicks.cs" />
     <Compile Include="Sections\Reports\EmailActivity\EmailActivityCollectionFilter.cs" />
     <Compile Include="Sections\Reports\EmailActivity\EmailActivityManager.cs" />
     <Compile Include="Sections\Reports\EmailActivity\EmailActivityReport.cs" />
     <Compile Include="Sections\Reports\EmailActivity\EmailActivityReportCollection.cs" />
     <Compile Include="Sections\Reports\EmailActivity\IEmailActivityManager.cs" />
+    <Compile Include="Sections\Reports\FacebookLikes.cs" />
+    <Compile Include="Sections\Reports\Forwards.cs" />
+    <Compile Include="Sections\Reports\IndustryStats.cs" />
     <Compile Include="Sections\Reports\IReportManager.cs" />
+    <Compile Include="Sections\Reports\ListStats.cs" />
+    <Compile Include="Sections\Reports\Opens.cs" />
     <Compile Include="Sections\Reports\Report.cs" />
     <Compile Include="Sections\Reports\ReportCollection.cs" />
     <Compile Include="Sections\Reports\ReportFilter.cs" />
     <Compile Include="Sections\Reports\ReportManager.cs" />
+    <Compile Include="Sections\Reports\ShareReport.cs" />
+    <Compile Include="Sections\Reports\TimeSeries.cs" />
+    <Compile Include="Sections\Reports\Timewarp.cs" />
   </ItemGroup>
   <ItemGroup>
     <Analyzer Include="..\packages\StyleCop.Analyzers.1.0.0\analyzers\dotnet\cs\Newtonsoft.Json.dll" />
