+++ MailChimpV3API/Sections/Reports/ShareReport.cs
@@ -1,0 +1,13 @@
+using Newtonsoft.Json;
+
+namespace MailChimpV3API.Sections.Reports
+{
+    public class ShareReport
+    {
+        [JsonProperty("share_url")]
+        public string ShareURL { get; set; }
+
+        [JsonProperty("share_password")]
+        public string SharePassword { get; set; }
+    }
+}
